Refactor 4 optimizer rules into separate files
