+++ arangod/Aql/OptimizerRules.cpp
@@ -26,10 +26,8 @@
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Aql/Ast.h"
-#include "Aql/AqlFunctionFeature.h"
 #include "Aql/AttributeNamePath.h"
 #include "Aql/Collection.h"
-#include "Aql/ConditionFinder.h"
 #include "Aql/ExecutionEngine.h"
 #include "Aql/ExecutionNode/CalculationNode.h"
 #include "Aql/ExecutionNode/DistributeNode.h"
@@ -58,13 +56,11 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionNode/TraversalNode.h"
 #include "Aql/ExecutionNode/MaterializeNode.h"
-#include "Aql/ExecutionNode/UpdateNode.h"
 #include "Aql/ExecutionNode/UpsertNode.h"
 #include "Aql/ExecutionNode/WindowNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
-#include "Aql/IndexHint.h"
 #include "Aql/Optimizer.h"
 #include "Aql/Optimizer/Utils/VariableReplacer.h"
 #include "Aql/OptimizerUtils.h"
@@ -74,16 +70,12 @@
 #include "Aql/SortElement.h"
 #include "Aql/SortInformation.h"
 #include "Aql/TraversalConditionFinder.h"
-#include "Aql/TypedAstNodes.h"
 #include "Aql/Variable.h"
 #include "Aql/types.h"
-#include "Basics/AttributeNameParser.h"
 #include "Basics/StaticStrings.h"
-#include "Basics/SupervisedBuffer.h"
 #include "Containers/FlatHashSet.h"
 #include "Containers/SmallUnorderedMap.h"
 #include "Containers/SmallVector.h"
-#include "Geo/GeoParams.h"
 #include "Graph/ShortestPathOptions.h"
 #include "Indexes/Index.h"
 #include "StorageEngine/StorageEngine.h"
@@ -130,13 +122,6 @@ static constexpr std::initializer_list<arangodb::aql::ExecutionNode::NodeType>
         arangodb::aql::ExecutionNode::UPDATE,
         arangodb::aql::ExecutionNode::REPLACE,
         arangodb::aql::ExecutionNode::REMOVE,
-        arangodb::aql::ExecutionNode::UPSERT};
-static constexpr std::initializer_list<arangodb::aql::ExecutionNode::NodeType>
-    removeDataModificationOutVariablesNodeTypes{
-        arangodb::aql::ExecutionNode::REMOVE,
-        arangodb::aql::ExecutionNode::INSERT,
-        arangodb::aql::ExecutionNode::UPDATE,
-        arangodb::aql::ExecutionNode::REPLACE,
         arangodb::aql::ExecutionNode::UPSERT};
 }  // namespace
 
@@ -735,1201 +720,6 @@ auto arangodb::aql::insertDistributeGatherSnippet(ExecutionPlan& plan,
     parents[0]->replaceDependency(deps[0], gatherNode);
   }
   return ExecutionNode::castTo<DistributeNode*>(distNode);
-}
-
-/// @brief remove $OLD and $NEW variables from data-modification statements
-/// if not required
-void arangodb::aql::removeDataModificationOutVariablesRule(
-    Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
-    OptimizerRule const& rule) {
-  bool modified = false;
-
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  plan->findNodesOfType(nodes, ::removeDataModificationOutVariablesNodeTypes,
-                        true);
-
-  for (auto const& n : nodes) {
-    auto node = ExecutionNode::castTo<ModificationNode*>(n);
-    TRI_ASSERT(node != nullptr);
-
-    Variable const* old = node->getOutVariableOld();
-    if (!n->isVarUsedLater(old)) {
-      // "$OLD" is not used later
-      node->clearOutVariableOld();
-      modified = true;
-    } else {
-      switch (n->getType()) {
-        case EN::UPDATE:
-        case EN::REPLACE: {
-          Variable const* inVariable =
-              ExecutionNode::castTo<UpdateReplaceNode const*>(n)
-                  ->inKeyVariable();
-          if (inVariable != nullptr) {
-            auto setter = plan->getVarSetBy(inVariable->id);
-            if (setter != nullptr &&
-                (setter->getType() == EN::ENUMERATE_COLLECTION ||
-                 setter->getType() == EN::INDEX)) {
-              std::unordered_map<VariableId, Variable const*> replacements;
-              replacements.try_emplace(old->id, inVariable);
-              VariableReplacer finder(replacements);
-              plan->root()->walk(finder);
-              modified = true;
-            }
-          }
-          break;
-        }
-        case EN::REMOVE: {
-          Variable const* inVariable =
-              ExecutionNode::castTo<RemoveNode const*>(n)->inVariable();
-          TRI_ASSERT(inVariable != nullptr);
-          auto setter = plan->getVarSetBy(inVariable->id);
-          if (setter != nullptr &&
-              (setter->getType() == EN::ENUMERATE_COLLECTION ||
-               setter->getType() == EN::INDEX)) {
-            std::unordered_map<VariableId, Variable const*> replacements;
-            replacements.try_emplace(old->id, inVariable);
-            VariableReplacer finder(replacements);
-            plan->root()->walk(finder);
-            modified = true;
-          }
-          break;
-        }
-        default: {
-          // do nothing
-        }
-      }
-    }
-
-    if (!n->isVarUsedLater(node->getOutVariableNew())) {
-      // "$NEW" is not used later
-      node->clearOutVariableNew();
-      modified = true;
-    }
-
-    if (!n->hasParent()) {
-      node->producesResults(false);
-      modified = true;
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, modified);
-}
-
-/// @brief optimizes away unused K_PATHS things
-void arangodb::aql::optimizePathsRule(Optimizer* opt,
-                                      std::unique_ptr<ExecutionPlan> plan,
-                                      OptimizerRule const& rule) {
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  plan->findNodesOfType(nodes, EN::ENUMERATE_PATHS, true);
-
-  if (nodes.empty()) {
-    // no traversals present
-    opt->addPlan(std::move(plan), rule, false);
-    return;
-  }
-
-  bool modified = false;
-
-  // first make a pass over all traversal nodes and remove unused
-  // variables from them
-  for (auto const& n : nodes) {
-    EnumeratePathsNode* ksp = ExecutionNode::castTo<EnumeratePathsNode*>(n);
-    if (!ksp->usesPathOutVariable()) {
-      continue;
-    }
-
-    Variable const* variable = &(ksp->pathOutVariable());
-
-    containers::FlatHashSet<AttributeNamePath> attributes;
-    VarSet vars;
-    bool canOptimize = true;
-
-    ExecutionNode* current = ksp->getFirstParent();
-    arangodb::ResourceMonitor& resMonitor =
-        plan->getAst()->query().resourceMonitor();
-
-    while (current != nullptr && canOptimize) {
-      switch (current->getType()) {
-        case EN::CALCULATION: {
-          vars.clear();
-          current->getVariablesUsedHere(vars);
-          if (vars.find(variable) != vars.end()) {
-            // path variable used here
-            Expression* exp =
-                ExecutionNode::castTo<CalculationNode*>(current)->expression();
-            AstNode const* node = exp->node();
-            if (!Ast::getReferencedAttributesRecursive(
-                    node, variable, /*expectedAttributes*/ "", attributes,
-                    resMonitor)) {
-              // full path variable is used, or accessed in a way that we don't
-              // understand, e.g. "p" or "p[0]" or "p[*]..."
-              canOptimize = false;
-            }
-          }
-          break;
-        }
-        default: {
-          // if the path is used by any other node type, we don't know what to
-          // do and will not optimize parts of it away
-          vars.clear();
-          current->getVariablesUsedHere(vars);
-          if (vars.find(variable) != vars.end()) {
-            canOptimize = false;
-          }
-          break;
-        }
-      }
-      current = current->getFirstParent();
-    }
-
-    if (canOptimize) {
-      bool produceVertices =
-          attributes.contains({StaticStrings::GraphQueryVertices, resMonitor});
-
-      if (!produceVertices) {
-        auto* options = ksp->options();
-        options->setProduceVertices(false);
-        modified = true;
-      }
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, modified);
-}
-
-/// Essentially mirrors the geo::QueryParams struct, but with
-/// abstracts AstNode value objects
-struct GeoIndexInfo {
-  operator bool() const {
-    return collectionNodeToReplace != nullptr && collectionNodeOutVar &&
-           collection && index && valid;
-  }
-  void invalidate() { valid = false; }
-
-  /// node that will be replaced by (geo) IndexNode
-  ExecutionNode* collectionNodeToReplace = nullptr;
-  Variable const* collectionNodeOutVar = nullptr;
-
-  /// accessed collection
-  aql::Collection const* collection = nullptr;
-  /// selected index
-  std::shared_ptr<Index> index;
-
-  /// Filter calculations to modify
-  std::map<ExecutionNode*, Expression*> exesToModify;
-  std::set<AstNode const*> nodesToRemove;
-
-  // ============ Distance ============
-  AstNode const* distCenterExpr = nullptr;
-  AstNode const* distCenterLatExpr = nullptr;
-  AstNode const* distCenterLngExpr = nullptr;
-  // Expression representing minimum distance
-  AstNode const* minDistanceExpr = nullptr;
-  // Was operator < or <= used
-  bool minInclusive = true;
-  // Expression representing maximum distance
-  AstNode const* maxDistanceExpr = nullptr;
-  // Was operator > or >= used
-  bool maxInclusive = true;
-
-  // ============ Near Info ============
-  bool sorted = false;
-  /// Default order is from closest to farthest
-  bool ascending = true;
-
-  // ============ Filter Info ===========
-  geo::FilterType filterMode = geo::FilterType::NONE;
-  /// variable using the filter mask
-  AstNode const* filterExpr = nullptr;
-
-  // ============ Accessed Fields ============
-  AstNode const* locationVar = nullptr;   // access to location field
-  AstNode const* latitudeVar = nullptr;   // access path to latitude
-  AstNode const* longitudeVar = nullptr;  // access path to longitude
-
-  /// contains this node a valid condition
-  bool valid = true;
-};
-
-// checks 2 parameters of distance function if they represent a valid access to
-// latitude and longitude attribute of the geo index.
-// distance(a,b,c,d) - possible pairs are (a,b) and (c,d)
-static bool distanceFuncArgCheck(ExecutionPlan* plan, AstNode const* latArg,
-                                 AstNode const* lngArg, bool supportLegacy,
-                                 GeoIndexInfo& info) {
-  // note: this only modifies "info" if the function returns true
-  std::pair<Variable const*, std::vector<arangodb::basics::AttributeName>>
-      attributeAccess1;
-  std::pair<Variable const*, std::vector<arangodb::basics::AttributeName>>
-      attributeAccess2;
-  // first and second should be based on the same document - need to provide
-  // the document in order to see which collection is bound to it and if that
-  // collections supports geo-index
-  if (!latArg->isAttributeAccessForVariable(attributeAccess1, true) ||
-      !lngArg->isAttributeAccessForVariable(attributeAccess2, true)) {
-    return false;
-  }
-  TRI_ASSERT(attributeAccess1.first != nullptr);
-  TRI_ASSERT(attributeAccess2.first != nullptr);
-
-  ExecutionNode* setter1 = plan->getVarSetBy(attributeAccess1.first->id);
-  ExecutionNode* setter2 = plan->getVarSetBy(attributeAccess2.first->id);
-  if (setter1 == nullptr || setter1 != setter2 ||
-      setter1->getType() != EN::ENUMERATE_COLLECTION) {
-    return false;  // expect access of doc.lat, doc.lng or doc.loc[0],
-                   // doc.loc[1]
-  }
-
-  // get logical collection
-  auto collNode = ExecutionNode::castTo<EnumerateCollectionNode*>(setter1);
-  if (info.collectionNodeToReplace != nullptr &&
-      info.collectionNodeToReplace != collNode) {
-    return false;  // should probably never happen
-  }
-
-  // we should not access the LogicalCollection directly
-  auto indexes = collNode->collection()->indexes();
-  // check for suitiable indexes
-  for (std::shared_ptr<Index> idx : indexes) {
-    // check if current index is a geo-index
-    std::size_t fieldNum = idx->fields().size();
-    bool isGeo1 = idx->type() == Index::IndexType::TRI_IDX_TYPE_GEO1_INDEX &&
-                  supportLegacy;
-    bool isGeo2 = idx->type() == Index::IndexType::TRI_IDX_TYPE_GEO2_INDEX &&
-                  supportLegacy;
-    bool isGeo = idx->type() == Index::IndexType::TRI_IDX_TYPE_GEO_INDEX;
-
-    if ((isGeo2 || isGeo) && fieldNum == 2) {  // individual fields
-      // check access paths of attributes in ast and those in index match
-      if (idx->fields()[0] == attributeAccess1.second &&
-          idx->fields()[1] == attributeAccess2.second) {
-        if (info.index != nullptr && info.index != idx) {
-          return false;
-        }
-        info.index = idx;
-        info.latitudeVar = latArg;
-        info.longitudeVar = lngArg;
-        info.collectionNodeToReplace = collNode;
-        info.collectionNodeOutVar = collNode->outVariable();
-        info.collection = collNode->collection();
-        return true;
-      }
-    } else if ((isGeo1 || isGeo) && fieldNum == 1) {
-      std::vector<basics::AttributeName> fields1 = idx->fields()[0];
-      std::vector<basics::AttributeName> fields2 = idx->fields()[0];
-
-      velocypack::SupervisedBuffer sb(
-          plan->getAst()->query().resourceMonitor());
-      VPackBuilder builder(sb);
-      idx->toVelocyPack(builder, Index::makeFlags(Index::Serialize::Basics));
-      bool geoJson = basics::VelocyPackHelper::getBooleanValue(
-          builder.slice(), "geoJson", false);
-
-      fields1.back().name += geoJson ? "[1]" : "[0]";
-      fields2.back().name += geoJson ? "[0]" : "[1]";
-      if (fields1 == attributeAccess1.second &&
-          fields2 == attributeAccess2.second) {
-        if (info.index != nullptr && info.index != idx) {
-          return false;
-        }
-        info.index = idx;
-        info.latitudeVar = latArg;
-        info.longitudeVar = lngArg;
-        info.collectionNodeToReplace = collNode;
-        info.collectionNodeOutVar = collNode->outVariable();
-        info.collection = collNode->collection();
-        return true;
-      }
-    }  // if isGeo 1 or 2
-  }  // for index in collection
-  return false;
-}
-
-// checks parameter of GEO_* function
-static bool geoFuncArgCheck(ExecutionPlan* plan, AstNode const* args,
-                            bool supportLegacy, GeoIndexInfo& info) {
-  // note: this only modifies "info" if the function returns true
-  std::pair<Variable const*, std::vector<arangodb::basics::AttributeName>>
-      attributeAccess;
-  // "arg" is either `[doc.lat, doc.lng]` or `doc.geometry`
-  if (args->isArray() && args->numMembers() == 2) {
-    return distanceFuncArgCheck(plan, /*lat*/ args->getMemberUnchecked(1),
-                                /*lng*/ args->getMemberUnchecked(0),
-                                supportLegacy, info);
-  } else if (!args->isAttributeAccessForVariable(attributeAccess, true)) {
-    return false;  // no attribute access, no index check
-  }
-  TRI_ASSERT(attributeAccess.first != nullptr);
-  ExecutionNode* setter = plan->getVarSetBy(attributeAccess.first->id);
-  if (setter == nullptr || setter->getType() != EN::ENUMERATE_COLLECTION) {
-    return false;  // expected access of the for doc.attribute
-  }
-
-  // get logical collection
-  auto collNode = ExecutionNode::castTo<EnumerateCollectionNode*>(setter);
-  if (info.collectionNodeToReplace != nullptr &&
-      info.collectionNodeToReplace != collNode) {
-    return false;  // should probably never happen
-  }
-
-  // we should not access the LogicalCollection directly
-  auto indexes = collNode->collection()->indexes();
-  // check for suitiable indexes
-  for (std::shared_ptr<arangodb::Index> idx : indexes) {
-    // check if current index is a geo-index
-    bool isGeo =
-        idx->type() == arangodb::Index::IndexType::TRI_IDX_TYPE_GEO_INDEX;
-    if (isGeo && idx->fields().size() == 1) {  // individual fields
-      // check access paths of attributes in ast and those in index match
-      if (idx->fields()[0] == attributeAccess.second) {
-        if (info.index != nullptr && info.index != idx) {
-          return false;  // different index
-        }
-        info.index = idx;
-        info.locationVar = args;
-        info.collectionNodeToReplace = collNode;
-        info.collectionNodeOutVar = collNode->outVariable();
-        info.collection = collNode->collection();
-        return true;
-      }
-    }
-  }  // for index in collection
-  return false;
-}
-
-/// returns true if left side is same as right or lhs is null
-static bool isValidGeoArg(AstNode const* lhs, AstNode const* rhs) {
-  if (lhs == nullptr) {  // lhs is from the GeoIndexInfo struct
-    return true;         // if geoindex field is null everything is valid
-  } else if (lhs->type != rhs->type) {
-    return false;
-  } else if (lhs->isArray()) {  // expect `[doc.lng, doc.lat]`
-    if (lhs->numMembers() >= 2 && rhs->numMembers() >= 2) {
-      return isValidGeoArg(lhs->getMemberUnchecked(0),
-                           rhs->getMemberUnchecked(0)) &&
-             isValidGeoArg(lhs->getMemberUnchecked(1),
-                           rhs->getMemberUnchecked(1));
-    }
-    return false;
-  } else if (lhs->type == NODE_TYPE_REFERENCE) {
-    return static_cast<Variable const*>(lhs->getData())->id ==
-           static_cast<Variable const*>(rhs->getData())->id;
-  }
-  // compareAstNodes does not handle non const attribute access
-  std::pair<Variable const*, std::vector<arangodb::basics::AttributeName>> res1,
-      res2;
-  bool acc1 = lhs->isAttributeAccessForVariable(res1, true);
-  bool acc2 = rhs->isAttributeAccessForVariable(res2, true);
-  if (acc1 || acc2) {
-    return acc1 && acc2 && res1 == res2;  // same variable same path
-  }
-  return aql::compareAstNodes(lhs, rhs, false) == 0;
-}
-
-static bool checkDistanceFunc(ExecutionPlan* plan, AstNode const* funcNode,
-                              bool legacy, GeoIndexInfo& info) {
-  // note: this only modifies "info" if the function returns true
-  if (funcNode->type == NODE_TYPE_REFERENCE) {
-    // FOR x IN cc LET d = DISTANCE(...) FILTER d > 10 RETURN x
-    Variable const* var = static_cast<Variable const*>(funcNode->getData());
-    TRI_ASSERT(var != nullptr);
-    ExecutionNode* setter = plan->getVarSetBy(var->id);
-    if (setter == nullptr || setter->getType() != EN::CALCULATION) {
-      return false;
-    }
-    funcNode =
-        ExecutionNode::castTo<CalculationNode*>(setter)->expression()->node();
-  }
-  // get the ast node of the expression
-  if (!funcNode || funcNode->type != NODE_TYPE_FCALL ||
-      funcNode->numMembers() != 1) {
-    return false;
-  }
-  AstNode* fargs = funcNode->getMemberUnchecked(0);
-  auto func = static_cast<Function const*>(funcNode->getData());
-  if (fargs->numMembers() >= 4 &&
-      func->name == "DISTANCE") {  // allow DISTANCE(a,b,c,d)
-    if (info.distCenterExpr != nullptr) {
-      return false;  // do not allow mixing of DISTANCE and GEO_DISTANCE
-    }
-    if (isValidGeoArg(info.distCenterLatExpr, fargs->getMemberUnchecked(2)) &&
-        isValidGeoArg(info.distCenterLngExpr, fargs->getMemberUnchecked(3)) &&
-        distanceFuncArgCheck(plan, fargs->getMemberUnchecked(0),
-                             fargs->getMemberUnchecked(1), legacy, info)) {
-      info.distCenterLatExpr = fargs->getMemberUnchecked(2);
-      info.distCenterLngExpr = fargs->getMemberUnchecked(3);
-      return true;
-    } else if (isValidGeoArg(info.distCenterLatExpr,
-                             fargs->getMemberUnchecked(0)) &&
-               isValidGeoArg(info.distCenterLngExpr,
-                             fargs->getMemberUnchecked(1)) &&
-               distanceFuncArgCheck(plan, fargs->getMemberUnchecked(2),
-                                    fargs->getMemberUnchecked(3), legacy,
-                                    info)) {
-      info.distCenterLatExpr = fargs->getMemberUnchecked(0);
-      info.distCenterLngExpr = fargs->getMemberUnchecked(1);
-      return true;
-    }
-  } else if (fargs->numMembers() == 2 && func->name == "GEO_DISTANCE") {
-    if (info.distCenterLatExpr || info.distCenterLngExpr) {
-      return false;  // do not allow mixing of DISTANCE and GEO_DISTANCE
-    }
-    if (isValidGeoArg(info.distCenterExpr, fargs->getMemberUnchecked(1)) &&
-        geoFuncArgCheck(plan, fargs->getMemberUnchecked(0), legacy, info)) {
-      info.distCenterExpr = fargs->getMemberUnchecked(1);
-      return true;
-    } else if (isValidGeoArg(info.distCenterExpr,
-                             fargs->getMemberUnchecked(0)) &&
-               geoFuncArgCheck(plan, fargs->getMemberUnchecked(1), legacy,
-                               info)) {
-      info.distCenterExpr = fargs->getMemberUnchecked(0);
-      return true;
-    }
-  }
-  return false;
-}
-
-// contains the AstNode* a supported function?
-static bool checkGeoFilterFunction(ExecutionPlan* plan, AstNode const* funcNode,
-                                   GeoIndexInfo& info) {
-  // note: this only modifies "info" if the function returns true
-  // the expression must exist and it must be a function call
-  if (funcNode->type != NODE_TYPE_FCALL || funcNode->numMembers() != 1 ||
-      info.filterMode != geo::FilterType::NONE) {  // can't handle more than one
-    return false;
-  }
-
-  auto func = static_cast<Function const*>(funcNode->getData());
-  AstNode* fargs = funcNode->getMemberUnchecked(0);
-  bool contains = func->name == "GEO_CONTAINS";
-  bool intersect = func->name == "GEO_INTERSECTS";
-  if ((!contains && !intersect) || fargs->numMembers() != 2) {
-    return false;
-  }
-
-  AstNode* arg = fargs->getMemberUnchecked(1);
-  if (geoFuncArgCheck(plan, arg, /*legacy*/ true, info)) {
-    TRI_ASSERT(contains || intersect);
-    info.filterMode =
-        contains ? geo::FilterType::CONTAINS : geo::FilterType::INTERSECTS;
-    info.filterExpr = fargs->getMemberUnchecked(0);
-    TRI_ASSERT(info.index);
-    return true;
-  }
-  return false;
-}
-
-// checks if a node contanis a geo index function a valid operator
-// to use within a filter condition
-bool checkGeoFilterExpression(ExecutionPlan* plan, AstNode const* node,
-                              GeoIndexInfo& info) {
-  // checks @first `smaller` @second
-  // note: this only modifies "info" if the function returns true
-  auto eval = [&](AstNode const* first, AstNode const* second,
-                  bool lessequal) -> bool {
-    if (second->type == NODE_TYPE_VALUE &&  // only constants allowed
-        info.maxDistanceExpr == nullptr &&  // max distance is not yet set
-        checkDistanceFunc(plan, first, /*legacy*/ true, info)) {
-      TRI_ASSERT(info.index);
-      info.maxDistanceExpr = second;
-      info.maxInclusive = info.maxInclusive && lessequal;
-      info.nodesToRemove.insert(node);
-      return true;
-    } else if (first->type == NODE_TYPE_VALUE &&  // only constants allowed
-               info.minDistanceExpr ==
-                   nullptr &&  // min distance is not yet set
-               checkDistanceFunc(plan, second, /*legacy*/ true, info)) {
-      info.minDistanceExpr = first;
-      info.minInclusive = info.minInclusive && lessequal;
-      info.nodesToRemove.insert(node);
-      return true;
-    }
-    return false;
-  };
-
-  switch (node->type) {
-    case NODE_TYPE_FCALL:
-      if (checkGeoFilterFunction(plan, node, info)) {
-        info.nodesToRemove.insert(node);
-        return true;
-      }
-      return false;
-    // only DISTANCE is allowed with <=, <, >=, >
-    case NODE_TYPE_OPERATOR_BINARY_LE: {
-      ast::RelationalOperatorNode relOp(node);
-      return eval(relOp.getLeft(), relOp.getRight(), true);
-    }
-    case NODE_TYPE_OPERATOR_BINARY_LT: {
-      ast::RelationalOperatorNode relOp(node);
-      return eval(relOp.getLeft(), relOp.getRight(), false);
-    }
-    case NODE_TYPE_OPERATOR_BINARY_GE: {
-      ast::RelationalOperatorNode relOp(node);
-      return eval(relOp.getRight(), relOp.getLeft(), true);
-    }
-    case NODE_TYPE_OPERATOR_BINARY_GT: {
-      ast::RelationalOperatorNode relOp(node);
-      return eval(relOp.getRight(), relOp.getLeft(), false);
-    }
-    default:
-      return false;
-  }
-}
-
-static bool optimizeSortNode(ExecutionPlan* plan, SortNode* sort,
-                             GeoIndexInfo& info) {
-  // note: info will only be modified if the function returns true
-  TRI_ASSERT(sort->getType() == EN::SORT);
-  // we're looking for "SORT DISTANCE(x,y,a,b)"
-  SortElementVector const& elements = sort->elements();
-  if (elements.size() != 1) {  // can't do it
-    return false;
-  }
-  TRI_ASSERT(elements[0].var != nullptr);
-
-  // find the expression that is bound to the variable
-  // get the expression node that holds the calculation
-  ExecutionNode* setter = plan->getVarSetBy(elements[0].var->id);
-  if (setter == nullptr || setter->getType() != EN::CALCULATION) {
-    return false;  // setter could be enumerate list node e.g.
-  }
-  CalculationNode* calc = ExecutionNode::castTo<CalculationNode*>(setter);
-  Expression* expr = calc->expression();
-  if (expr == nullptr || expr->node() == nullptr) {
-    return false;  // the expression must exist and must have an astNode
-  }
-
-  // info will only be modified if the function returns true
-  bool legacy = elements[0].ascending;  // DESC is only supported on S2 index
-  if (!info.sorted && checkDistanceFunc(plan, expr->node(), legacy, info)) {
-    info.sorted = true;  // do not parse another SORT
-    info.ascending = elements[0].ascending;
-    if (!ServerState::instance()->isCoordinator()) {
-      // we must not remove a sort in the cluster... the results from each
-      // shard will be sorted by using the index, however we still need to
-      // establish a cross-shard sortedness by distance.
-      info.exesToModify.emplace(sort, expr);
-      info.nodesToRemove.emplace(expr->node());
-    } else {
-      // In the cluster case, we want to leave the SORT node in - for now!
-      // This is to achieve that the GATHER node which is introduced to
-      // distribute the query in the cluster remembers to sort things
-      // using merge sort. However, later there will be a rule which
-      // moves the sorting to the dbserver. When this rule is triggered,
-      // we do not want to reinsert the SORT node on the dbserver, since
-      // there, the items are already sorted by means of the geo index.
-      // Therefore, we tell the sort node here, not to be reinserted
-      // on the dbserver later on.
-      // This is crucial to avoid that the SORT node remains and pulls
-      // the whole collection out of the geo index, on the grounds that
-      // the SORT node wants to sort the results, which is very bad for
-      // performance.
-      sort->dontReinsertInCluster();
-    }
-    return true;
-  }
-  return false;
-}
-
-// checks a single sort or filter node
-static void optimizeFilterNode(ExecutionPlan* plan, FilterNode* fn,
-                               GeoIndexInfo& info) {
-  TRI_ASSERT(fn->getType() == EN::FILTER);
-
-  // filter nodes always have one input variable
-  auto variable = ExecutionNode::castTo<FilterNode const*>(fn)->inVariable();
-  // now check who introduced our variable
-  ExecutionNode* setter = plan->getVarSetBy(variable->id);
-  if (setter == nullptr || setter->getType() != EN::CALCULATION) {
-    return;
-  }
-  CalculationNode* calc = ExecutionNode::castTo<CalculationNode*>(setter);
-  Expression* expr = calc->expression();
-  if (expr->node() == nullptr) {
-    return;  // the expression must exist and must have an AstNode
-  }
-
-  Ast::traverseReadOnly(
-      expr->node(),
-      [&](AstNode const* node) {  // pre
-        if (node->isSimpleComparisonOperator() ||
-            node->type == arangodb::aql::NODE_TYPE_FCALL ||
-            node->type == arangodb::aql::NODE_TYPE_OPERATOR_BINARY_AND ||
-            node->type == arangodb::aql::NODE_TYPE_OPERATOR_NARY_AND) {
-          return true;
-        }
-        return false;
-      },
-      [&](AstNode const* node) {  // post
-        if (!node->isSimpleComparisonOperator() &&
-            node->type != arangodb::aql::NODE_TYPE_FCALL) {
-          return;
-        }
-        if (checkGeoFilterExpression(plan, node, info)) {
-          info.exesToModify.try_emplace(fn, expr);
-        }
-      });
-}
-
-// modify plan
-
-// builds a condition that can be used with the index interface and
-// contains all parameters required by the MMFilesGeoIndex
-static std::unique_ptr<Condition> buildGeoCondition(ExecutionPlan* plan,
-                                                    GeoIndexInfo const& info) {
-  Ast* ast = plan->getAst();
-  // shared code to add symbolic `doc.geometry` or `[doc.lng, doc.lat]`
-  auto addLocationArg = [ast, &info](AstNode* args) {
-    if (info.locationVar) {
-      args->addMember(info.locationVar);
-    } else if (info.latitudeVar && info.longitudeVar) {
-      AstNode* array = ast->createNodeArray(2);
-      array->addMember(info.longitudeVar);  // GeoJSON ordering
-      array->addMember(info.latitudeVar);
-      args->addMember(array);
-    } else {
-      TRI_ASSERT(false);
-      THROW_ARANGO_EXCEPTION_MESSAGE(TRI_ERROR_INTERNAL,
-                                     "unsupported geo type");
-    }
-  };
-
-  TRI_ASSERT(info.index);
-  auto cond = std::make_unique<Condition>(ast);
-  bool hasCenter = info.distCenterLatExpr || info.distCenterExpr;
-  bool hasDistLimit = info.maxDistanceExpr || info.minDistanceExpr;
-  TRI_ASSERT(!hasCenter || hasDistLimit || info.sorted);
-  if (hasCenter && (hasDistLimit || info.sorted)) {
-    // create GEO_DISTANCE(...) [<|<=|>=|>] Var
-    AstNode* args = ast->createNodeArray(2);
-    if (info.distCenterLatExpr && info.distCenterLngExpr) {  // legacy
-      TRI_ASSERT(!info.distCenterExpr);
-      // info.sorted && info.ascending &&
-      AstNode* array = ast->createNodeArray(2);
-      array->addMember(info.distCenterLngExpr);  // GeoJSON ordering
-      array->addMember(info.distCenterLatExpr);
-      args->addMember(array);
-    } else {
-      TRI_ASSERT(info.distCenterExpr);
-      TRI_ASSERT(!info.distCenterLatExpr && !info.distCenterLngExpr);
-      args->addMember(info.distCenterExpr);  // center location
-    }
-
-    addLocationArg(args);
-    AstNode* func = ast->createNodeFunctionCall("GEO_DISTANCE", args, true);
-
-    TRI_ASSERT(info.maxDistanceExpr || info.minDistanceExpr || info.sorted);
-    if (info.minDistanceExpr != nullptr) {
-      AstNodeType t = info.minInclusive ? NODE_TYPE_OPERATOR_BINARY_GE
-                                        : NODE_TYPE_OPERATOR_BINARY_GT;
-      cond->andCombine(
-          ast->createNodeBinaryOperator(t, func, info.minDistanceExpr));
-    }
-    if (info.maxDistanceExpr != nullptr) {
-      AstNodeType t = info.maxInclusive ? NODE_TYPE_OPERATOR_BINARY_LE
-                                        : NODE_TYPE_OPERATOR_BINARY_LT;
-      cond->andCombine(
-          ast->createNodeBinaryOperator(t, func, info.maxDistanceExpr));
-    }
-    if (info.minDistanceExpr == nullptr && info.maxDistanceExpr == nullptr &&
-        info.sorted) {
-      // hack to pass on the sort-to-point info
-      AstNodeType t = NODE_TYPE_OPERATOR_BINARY_LT;
-      std::string const& u = arangodb::StaticStrings::Unlimited;
-      AstNode* cc = ast->createNodeValueString(u.c_str(), u.length());
-      cond->andCombine(ast->createNodeBinaryOperator(t, func, cc));
-    }
-  }
-  if (info.filterMode != geo::FilterType::NONE) {
-    // create GEO_CONTAINS / GEO_INTERSECTS
-    TRI_ASSERT(info.filterExpr);
-    TRI_ASSERT(info.locationVar || (info.longitudeVar && info.latitudeVar));
-
-    AstNode* args = ast->createNodeArray(2);
-    args->addMember(info.filterExpr);
-    addLocationArg(args);
-    if (info.filterMode == geo::FilterType::CONTAINS) {
-      cond->andCombine(ast->createNodeFunctionCall("GEO_CONTAINS", args, true));
-    } else if (info.filterMode == geo::FilterType::INTERSECTS) {
-      cond->andCombine(
-          ast->createNodeFunctionCall("GEO_INTERSECTS", args, true));
-    } else {
-      TRI_ASSERT(false);
-    }
-  }
-
-  cond->normalize(plan);
-  return cond;
-}
-
-// applys the optimization for a candidate
-static bool applyGeoOptimization(ExecutionPlan* plan, LimitNode* ln,
-                                 GeoIndexInfo const& info) {
-  TRI_ASSERT(info.collection != nullptr);
-  TRI_ASSERT(info.collectionNodeToReplace != nullptr);
-  TRI_ASSERT(info.index);
-
-  // verify that all vars used in the index condition are valid
-  auto const& valid = info.collectionNodeToReplace->getVarsValid();
-  auto checkVars = [&valid](AstNode const* expr) {
-    if (expr != nullptr) {
-      VarSet varsUsed;
-      Ast::getReferencedVariables(expr, varsUsed);
-      for (Variable const* v : varsUsed) {
-        if (valid.find(v) == valid.end()) {
-          return false;  // invalid variable foud
-        }
-      }
-    }
-    return true;
-  };
-  if (!checkVars(info.distCenterExpr) || !checkVars(info.distCenterLatExpr) ||
-      !checkVars(info.distCenterLngExpr) || !checkVars(info.filterExpr)) {
-    return false;
-  }
-
-  size_t limit = 0;
-  if (ln != nullptr) {
-    limit = ln->offset() + ln->limit();
-    TRI_ASSERT(limit != SIZE_MAX);
-  }
-
-  IndexIteratorOptions opts;
-  opts.sorted = info.sorted;
-  opts.ascending = info.ascending;
-  opts.limit = limit;
-  opts.evaluateFCalls = false;  // workaround to avoid evaluating "doc.geo"
-  std::unique_ptr<Condition> condition(buildGeoCondition(plan, info));
-  auto inode = plan->createNode<IndexNode>(
-      plan, plan->nextId(), info.collection, info.collectionNodeOutVar,
-      std::vector<transaction::Methods::IndexHandle>{
-          transaction::Methods::IndexHandle{info.index}},
-      false,  // here we are not using inverted index so
-              // for sure no "whole" coverage
-      std::move(condition), opts);
-  plan->replaceNode(info.collectionNodeToReplace, inode);
-
-  // remove expressions covered by our index
-  Ast* ast = plan->getAst();
-  for (std::pair<ExecutionNode*, Expression*> pair : info.exesToModify) {
-    AstNode* root = pair.second->nodeForModification();
-    auto pre = [&](AstNode const* node) -> bool {
-      return node == root || Ast::isAndOperatorType(node->type);
-    };
-    auto visitor = [&](AstNode* node) -> AstNode* {
-      if (Ast::isAndOperatorType(node->type)) {
-        std::vector<AstNode*> keep;  // always shallow copy node
-        for (std::size_t i = 0; i < node->numMembers(); i++) {
-          AstNode* child = node->getMemberUnchecked(i);
-          if (info.nodesToRemove.find(child) == info.nodesToRemove.end()) {
-            keep.push_back(child);
-          }
-        }
-
-        if (keep.size() > 2) {
-          AstNode* n = ast->createNodeNaryOperator(NODE_TYPE_OPERATOR_NARY_AND);
-          for (size_t i = 0; i < keep.size(); i++) {
-            n->addMember(keep[i]);
-          }
-          return n;
-        } else if (keep.size() == 2) {
-          return ast->createNodeBinaryOperator(NODE_TYPE_OPERATOR_BINARY_AND,
-                                               keep[0], keep[1]);
-        } else if (keep.size() == 1) {
-          return keep[0];
-        }
-        return node == root ? nullptr : ast->createNodeValueBool(true);
-      } else if (info.nodesToRemove.find(node) != info.nodesToRemove.end()) {
-        return node == root ? nullptr : ast->createNodeValueBool(true);
-      }
-      return node;
-    };
-    auto post = [](AstNode const*) {};
-    AstNode* newNode = Ast::traverseAndModify(root, pre, visitor, post);
-    if (newNode == nullptr) {  // if root was removed, unlink FILTER or SORT
-      plan->unlinkNode(pair.first);
-    } else if (newNode != root) {
-      pair.second->replaceNode(newNode);
-    }
-  }
-
-  // signal that plan has been changed
-  return true;
-}
-
-void arangodb::aql::geoIndexRule(Optimizer* opt,
-                                 std::unique_ptr<ExecutionPlan> plan,
-                                 OptimizerRule const& rule) {
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  bool mod = false;
-
-  plan->findNodesOfType(nodes, EN::ENUMERATE_COLLECTION, true);
-  for (ExecutionNode* node : nodes) {
-    GeoIndexInfo info;
-    info.collectionNodeToReplace = node;
-
-    ExecutionNode* current = node->getFirstParent();
-    LimitNode* limit = nullptr;
-    bool canUseSortLimit = true;
-    bool mustRespectIdxHint = false;
-    auto enumerateColNode =
-        ExecutionNode::castTo<EnumerateCollectionNode const*>(node);
-    auto const& colNodeHints = enumerateColNode->hint();
-    if (colNodeHints.isForced() && colNodeHints.isSimple()) {
-      auto indexes = enumerateColNode->collection()->indexes();
-      auto& idxNames = colNodeHints.candidateIndexes();
-      for (auto const& idxName : idxNames) {
-        for (std::shared_ptr<Index> const& idx : indexes) {
-          if (idx->name() == idxName) {
-            auto idxType = idx->type();
-            if ((idxType != Index::IndexType::TRI_IDX_TYPE_GEO1_INDEX) &&
-                (idxType != Index::IndexType::TRI_IDX_TYPE_GEO2_INDEX) &&
-                (idxType != Index::IndexType::TRI_IDX_TYPE_GEO_INDEX)) {
-              mustRespectIdxHint = true;
-            } else {
-              info.index = idx;
-            }
-            break;
-          }
-        }
-      }
-    }
-
-    while (current) {
-      if (current->getType() == EN::FILTER) {
-        // picking up filter conditions is always allowed
-        optimizeFilterNode(plan.get(),
-                           ExecutionNode::castTo<FilterNode*>(current), info);
-      } else if (current->getType() == EN::SORT && canUseSortLimit) {
-        // only pick up a sort clause if we haven't seen another loop yet
-        if (!optimizeSortNode(
-                plan.get(), ExecutionNode::castTo<SortNode*>(current), info)) {
-          // 1. EnumerateCollectionNode x
-          // 2. SortNode x.abc ASC
-          // 3. LimitNode n,m  <-- cannot reuse LIMIT node here
-          // limit = nullptr;
-          break;  // stop parsing on non-optimizable SORT
-        }
-      } else if (current->getType() == EN::LIMIT && canUseSortLimit) {
-        // only pick up a limit clause if we haven't seen another loop yet
-        limit = ExecutionNode::castTo<LimitNode*>(current);
-        break;  // stop parsing after first LIMIT
-      } else if (current->getType() == EN::RETURN ||
-                 current->getType() == EN::COLLECT) {
-        break;  // stop parsing on return or collect
-      } else if (current->getType() == EN::INDEX ||
-                 current->getType() == EN::ENUMERATE_COLLECTION ||
-                 current->getType() == EN::ENUMERATE_LIST ||
-                 current->getType() == EN::ENUMERATE_IRESEARCH_VIEW ||
-                 current->getType() == EN::TRAVERSAL ||
-                 current->getType() == EN::ENUMERATE_PATHS ||
-                 current->getType() == EN::SHORTEST_PATH) {
-        // invalidate limit and sort. filters can still be used
-        limit = nullptr;
-        info.sorted = false;
-        // don't allow picking up either sort or limit from here on
-        canUseSortLimit = false;
-      }
-      current = current->getFirstParent();  // inspect next node
-    }
-
-    // if info is valid we try to optimize ENUMERATE_COLLECTION
-    if (info && info.collectionNodeToReplace == node) {
-      if (!mustRespectIdxHint &&
-          applyGeoOptimization(plan.get(), limit, info)) {
-        mod = true;
-      }
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, mod);
-}
-
-/// @brief turn LENGTH(FOR doc IN ...) subqueries into an optimized count
-/// operation
-void arangodb::aql::optimizeCountRule(Optimizer* opt,
-                                      std::unique_ptr<ExecutionPlan> plan,
-                                      OptimizerRule const& rule) {
-  bool modified = false;
-
-  if (plan->getAst()->query().queryOptions().fullCount) {
-    // fullCount is unsupported yet
-    opt->addPlan(std::move(plan), rule, modified);
-    return;
-  }
-
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  plan->findNodesOfType(nodes, EN::CALCULATION, true);
-
-  VarSet vars;
-  std::unordered_map<ExecutionNode*,
-                     std::pair<bool, std::unordered_set<AstNode const*>>>
-      candidates;
-
-  // find all calculation nodes in the plan
-  for (auto const& n : nodes) {
-    auto cn = ExecutionNode::castTo<CalculationNode*>(n);
-    AstNode const* root = cn->expression()->node();
-    if (root == nullptr) {
-      continue;
-    }
-
-    std::unordered_map<ExecutionNode*,
-                       std::pair<bool, std::unordered_set<AstNode const*>>>
-        localCandidates;
-
-    // look for all expressions that contain COUNT(subquery) or LENGTH(subquery)
-    auto visitor = [&localCandidates, &plan](AstNode const* node) -> bool {
-      if (node->type == NODE_TYPE_FCALL && node->numMembers() > 0) {
-        ast::FunctionCallNode fcall(node);
-        auto func = fcall.getFunction();
-        auto args = fcall.getArguments().getElements();
-        if (func->name == "LENGTH" || func->name == "COUNT") {
-          if (args.size() > 0 && args[0]->type == NODE_TYPE_REFERENCE) {
-            ast::ReferenceNode ref(args[0]);
-            Variable const* v = ref.getVariable();
-            auto setter = plan->getVarSetBy(v->id);
-            if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
-              // COUNT(subquery) / LENGTH(subquery)
-              auto sn = ExecutionNode::castTo<SubqueryNode*>(setter);
-              if (sn->isModificationNode()) {
-                // subquery modifies data
-                // cannot apply optimization for data-modification queries
-                return true;
-              }
-              if (!sn->isDeterministic()) {
-                // subquery is non-deterministic. cannot apply the optimization
-                return true;
-              }
-
-              auto current = sn->getSubquery();
-              if (current == nullptr || current->getType() != EN::RETURN) {
-                // subquery does not end with a RETURN instruction - we cannot
-                // handle this
-                return true;
-              }
-
-              auto it = localCandidates.find(setter);
-              if (it == localCandidates.end()) {
-                localCandidates.emplace(
-                    setter,
-                    std::make_pair(true,
-                                   std::unordered_set<AstNode const*>({node})));
-              } else {
-                (*it).second.second.emplace(node);
-              }
-              return false;
-            }
-          }
-        }
-      } else if (node->type == NODE_TYPE_REFERENCE) {
-        Variable const* v = static_cast<Variable const*>(node->getData());
-        auto setter = plan->getVarSetBy(v->id);
-        if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
-          // subquery used for something else inside the calculation,
-          // e.g. FIRST(subquery).
-          // we cannot continue with the optimization for this subquery, but for
-          // others
-          localCandidates[setter].first = false;
-          return false;
-        }
-      }
-      return true;
-    };
-
-    Ast::traverseReadOnly(root, visitor, [](AstNode const*) {});
-
-    for (auto const& it : localCandidates) {
-      // check if subquery result is used for something else than LENGTH/COUNT
-      // in *this* calculation
-      if (!it.second.first) {
-        // subquery result is used for other calculations than COUNT(subquery)
-        continue;
-      }
-
-      SubqueryNode const* sn =
-          ExecutionNode::castTo<SubqueryNode const*>(it.first);
-      if (n->isVarUsedLater(sn->outVariable())) {
-        // subquery result is used elsewhere later - we cannot optimize
-        continue;
-      }
-
-      bool valid = true;
-      // check if subquery result is used somewhere else before the current
-      // calculation we are looking at
-      auto current = sn->getFirstParent();
-      while (current != nullptr && current != n) {
-        vars.clear();
-        current->getVariablesUsedHere(vars);
-        if (vars.find(sn->outVariable()) != vars.end()) {
-          valid = false;
-          break;
-        }
-        current = current->getFirstParent();
-      }
-
-      if (valid) {
-        // subquery result is not used elsewhere - we can continue optimizing
-        // transfer the candidate into the global result
-        candidates.emplace(it.first, it.second);
-      }
-    }
-  }
-
-  for (auto const& it : candidates) {
-    TRI_ASSERT(it.second.first);
-    auto sn = ExecutionNode::castTo<SubqueryNode const*>(it.first);
-
-    // scan from the subquery node to the bottom of the ExecutionPlan to check
-    // if any of the following nodes also use the subquery result
-    auto current = sn->getSubquery();
-    TRI_ASSERT(current->getType() == EN::RETURN);
-    auto returnNode = ExecutionNode::castTo<ReturnNode*>(current);
-    auto returnSetter = plan->getVarSetBy(returnNode->inVariable()->id);
-    if (returnSetter == nullptr) {
-      continue;
-    }
-    if (returnSetter->getType() == EN::CALCULATION) {
-      // check if we can understand this type of calculation
-      auto cn = ExecutionNode::castTo<CalculationNode*>(returnSetter);
-      auto expr = cn->expression();
-      if (!expr->isConstant() && !expr->isAttributeAccess()) {
-        continue;
-      }
-    }
-
-    // find the head of the plan/subquery
-    while (current->hasDependency()) {
-      current = current->getFirstDependency();
-    }
-
-    TRI_ASSERT(current != nullptr);
-
-    if (current->getType() != EN::SINGLETON) {
-      continue;
-    }
-
-    // from here we need to find the first FOR loop.
-    // if it is a full collection scan or an index scan, we note its out
-    // variable. if we find a nested loop, we abort searching
-    bool valid = true;
-    ExecutionNode* found = nullptr;
-    Variable const* outVariable = nullptr;
-    current = current->getFirstParent();
-
-    while (current != nullptr) {
-      auto type = current->getType();
-      switch (type) {
-        case EN::ENUMERATE_COLLECTION:
-        case EN::INDEX: {
-          if (found != nullptr) {
-            // found a nested collection/index scan
-            found = nullptr;
-            valid = false;
-          } else {
-            TRI_ASSERT(valid);
-            if (dynamic_cast<DocumentProducingNode*>(current)->hasFilter()) {
-              // node uses early pruning. this is not supported
-              valid = false;
-            } else {
-              outVariable =
-                  dynamic_cast<DocumentProducingNode*>(current)->outVariable();
-
-              if (type == EN::INDEX &&
-                  ExecutionNode::castTo<IndexNode const*>(current)
-                          ->getIndexes()
-                          .size() != 1) {
-                // more than one index, so we would need to run uniqueness
-                // checks on the results. this is currently unsupported, so
-                // don't apply the optimization
-                valid = false;
-              } else {
-                // a FOR loop without an early pruning filter. this is what we
-                // are looking for!
-                found = current;
-              }
-            }
-          }
-          break;
-        }
-
-        case EN::DISTRIBUTE:
-
-        case EN::INSERT:
-        case EN::UPDATE:
-        case EN::REPLACE:
-        case EN::REMOVE:
-        case EN::UPSERT:
-          // we don't handle data-modification queries
-
-        case EN::LIMIT:
-          // limit is not yet supported
-
-        case EN::ENUMERATE_LIST:
-        case EN::TRAVERSAL:
-        case EN::SHORTEST_PATH:
-        case EN::ENUMERATE_PATHS:
-        case EN::ENUMERATE_IRESEARCH_VIEW: {
-          // we don't handle nested FOR loops
-          found = nullptr;
-          valid = false;
-          break;
-        }
-
-        case EN::RETURN: {
-          // we reached the end
-          break;
-        }
-
-        default: {
-          if (outVariable != nullptr) {
-            vars.clear();
-            current->getVariablesUsedHere(vars);
-            if (vars.find(outVariable) != vars.end()) {
-              // result variable of FOR loop is used somewhere where we
-              // can't handle it - don't apply the optimization
-              found = nullptr;
-              valid = false;
-            }
-          }
-          break;
-        }
-      }
-
-      if (!valid) {
-        break;
-      }
-
-      current = current->getFirstParent();
-    }
-
-    if (valid && found != nullptr) {
-      dynamic_cast<DocumentProducingNode*>(found)->setCountFlag();
-      returnNode->inVariable(outVariable);
-
-      // replace COUNT/LENGTH with SUM, as we are getting an array from the
-      // subquery
-      auto& server = plan->getAst()->query().vocbase().server();
-      auto func = server.getFeature<AqlFunctionFeature>().byName("SUM");
-      for (AstNode const* funcNode : it.second.second) {
-        const_cast<AstNode*>(funcNode)->setData(static_cast<void const*>(func));
-      }
-
-      if (returnSetter->getType() == EN::CALCULATION) {
-        plan->clearVarUsageComputed();
-        plan->findVarUsage();
-
-        auto cn = ExecutionNode::castTo<CalculationNode*>(returnSetter);
-        if (cn->expression()->isConstant() &&
-            !cn->isVarUsedLater(cn->outVariable())) {
-          plan->unlinkNode(cn);
-        }
-      }
-      modified = true;
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, modified);
 }
 
 void arangodb::aql::asyncPrefetchRule(Optimizer* opt,
