+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -55,6 +55,10 @@
 #include "Aql/Optimizer/Rule/SortInValues.h"
 #include "Aql/Optimizer/Rule/SpecializeCollect.h"
 #include "Aql/Optimizer/Rule/SpliceSubqueries.h"
+#include "Aql/Optimizer/Rule/RemoveDataModificationOutVariables.h"
+#include "Aql/Optimizer/Rule/OptimizePaths.h"
+#include "Aql/Optimizer/Rule/GeoIndex.h"
+#include "Aql/Optimizer/Rule/OptimizeCount.h"
 #include "Aql/Optimizer/Rule/DistributeInCluster.h"
 #include "Aql/Optimizer/Rule/CollectInCluster.h"
 #include "Aql/Optimizer/Rule/DistributeFilterCalcToCluster.h"
